+++ analytics/onetrust.ejs
@@ -1,0 +1,12 @@
+<%# Also used in views/welcome to enable onetrust in the welcome page %>
+<% if (locals.analytics.onetrust) { %>
+  <!-- OneTrust cookie consent banner -->
+  <script
+    id="cookie-consent-script"
+    charset="UTF-8"
+    type="text/javascript"
+    data-domain-script="<%= locals.analytics.onetrust.domainId %>"
+    src="https://cdn.cookielaw.org/scripttemplates/otSDKStub.js"
+  ></script>
+  <!-- End OneTrust cookie consent banner -->
+<% } %>
